Add rules for nethack-qt target
In addition to rules corresponding to those for the other packages, this adds
the LINK variable for all packages.  The Qt interface needs to be linked with
g++ instead of cc or gcc.  The make file (sys/unix/Makefile.src) already uses
this variable, and did not need to be modified.